Add zero-allocation test vector benchmarks and libopus comparison

+++ pcm.go
@@ -3,14 +3,14 @@ package gopus
 import "math"
 
 func float32ToInt16(sample float32) int16 {
-	scaled := float64(sample) * 32768.0
+	scaled := sample * 32768.0
 	if scaled > 32767.0 {
 		return 32767
 	}
 	if scaled < -32768.0 {
 		return -32768
 	}
-	return int16(math.RoundToEven(scaled))
+	return int16(math.RoundToEven(float64(scaled)))
 }
 
 func float64ToInt16(sample float64) int16 {
